// src/types.ts
export type Location = {
    id: number;
    name: string;
    latitude: number;
    longitude: number;
    elevation: number;
    feature_code: string;
    country_code: string;
    timezone: string;
    population: number;
    postcodes: string[];
    country_id: number;
    country: string;
    admin1?: string;
    admin2?: string;
    admin3?: string;
    admin4?: string;
    admin1_id?: string;
    admin2_id?: string;
    admin3_id?: string;
    admin4_id?: string;
}

export type LocationResponse = {
    results?: Location[];
    generationtime_ms: number;
}

export type WeatherResponse = {
    latitude: number,
    longitude: number,
    generationtime_ms: number,
    utc_offset_seconds: number,
    timezone: string,
    timezone_abbreviation: string,
    elevation: number,
    current_weather_units: {
        time: string,
        interval: string,
        temperature: string,
        windspeed: string,
        winddirection: string,
        is_day: string,
        weathercode: string,
    },
    current_weather: {
        time: string,
        interval: number,
        temperature: number,
        windspeed: number,
        winddirection: number,
        is_day: number,
        weathercode: number
    }
}